backend/requirements.txt
# FastAPI and async
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP client for agent calls
httpx==0.25.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Code quality
black==23.12.0
flake8==6.1.0
isort==5.13.2
mypy==1.7.1

# Utilities
python-dotenv==1.0.0
uuid==1.30
